+++ src/apps/gamma-strategies/celo/gamma-strategies.pool.token-fetcher.ts
@@ -1,7 +1,8 @@
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
 
-import { GammaStrategiesPoolTokenFetcher } from '../helpers/gamma-strategies.pool.token-fetcher'
+import { GammaStrategiesPoolTokenFetcher } from '../common/gamma-strategies.pool.token-fetcher';
 
 @PositionTemplate()
 export class CeloGammaStrategiesPoolTokenFetcher extends GammaStrategiesPoolTokenFetcher {
+  groupLabel = 'Pools';
 }