+++ src/apps/kwenta/common/kwenta.perp.contract-position-fetcher.ts
@@ -13,6 +13,7 @@ import {
   GetTokenBalancesParams,
 } from '~position/template/contract-position.template.types';
 
+import { PERP_V2_MARKET_ADDRESSES } from '../../synthetix/optimism/synthetix.perp-v2.contract-position-fetcher';
 import { KwentaContractFactory, KwentaFutures } from '../contracts';
 
 type GetContracts = {
@@ -47,7 +48,12 @@ export abstract class OptimismKwentaPerpContractPositionFetcher extends Contract
       query: getContractsQuery,
     });
 
-    return contractsFromSubgraph.futuresMarkets.map(futuresMarket => ({ address: futuresMarket.id }));
+    return contractsFromSubgraph.futuresMarkets
+      .filter(market => {
+        const isPerpV2Market = PERP_V2_MARKET_ADDRESSES.find(element => element.address === market.id);
+        return !isPerpV2Market;
+      })
+      .map(futuresMarket => ({ address: futuresMarket.id }));
   }
 
   async getTokenDefinitions() {