+++ src/apps/unipilot/unipilot.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 
 import { AbstractApp } from '~app/app.dynamic-module';
 
-import { UnipilotVaultAPYHelper } from './common/unipilot-vault.apy.helper';
 import { UnipilotVaultDefinitionsResolver } from './common/unipilot.vault-definition-resolver';
 import { UnipilotContractFactory } from './contracts';
 import { EthereumUnipilotPoolTokenFetcher } from './ethereum/unipilot.pool.token-fetcher';
@@ -12,7 +11,6 @@ import { PolygonUnipilotPoolTokenFetcher } from './polygon/unipilot.pool.token-f
   providers: [
     UnipilotContractFactory,
     UnipilotVaultDefinitionsResolver,
-    UnipilotVaultAPYHelper,
     // Ethereum
     EthereumUnipilotPoolTokenFetcher,
     // Polygon